feat: prepare AnnexThemeField
